Mod: Changed the Month Format in formatDate
Shortened the outputted month text ex:
December -> Dec
January -> Jan

+++ InternalWeb/.JS/DueTimeCalculator.js
@@ -8,6 +8,6 @@ function getDueTime(dueDate) {
 
 function formatDate(inputDate) {
     const date = new Date(inputDate);
-    const options = { year: 'numeric', month: 'long', day: 'numeric' };
+    const options = { year: 'numeric', month: 'short', day: 'numeric' };
     return date.toLocaleDateString('en-US', options);
 }
